% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/util_acceptance_rate.R
\name{acceptance_rate}
\alias{acceptance_rate}
\title{Compute acceptance rate}
\usage{
acceptance_rate(...)
}
\arguments{
\item{...}{parameters to \link{\code{get_traces}} (especially 'run')}
}
\value{
acceptance rate
}
\description{
This function takes the provided \code{bi_wrapper} which has been
run, or a bi file, and returns a the acceptance rate
}

