% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/util_adapt_mcmc.R
\name{adapt_mcmc}
\alias{adapt_mcmc}
\title{Adapt the proposal distribution of MCMC using the covariance of samples}
\usage{
adapt_mcmc(wrapper, min = 0, max = 1, scale, add_options, samples,
  max_iter = 10, ...)
}
\arguments{
\item{wrapper}{\code{bi_wrapper} (which has been run) to study}

\item{min}{minimum acceptance rate}

\item{max}{maximum acceptance rate}

\item{scale}{scale multiplier for the proposal}

\item{add_options}{list of additional options}

\item{samples}{number of samples to generate each iteration}

\item{max_iter}{maximum of iterations}

\item{...}{parameters for bi_wrapper$run}
}
\value{
a \code{bi_wrapper} with the desired proposal distribution
}
\description{
This function takes the provided \code{bi_wrapper} and runs MCMC,
adapting the proposal distribution until the desired acceptance
rate is achieved
}

