% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/util_adapt_particles.R
\name{adapt_particles}
\alias{adapt_particles}
\title{Adapt the number of particles}
\usage{
adapt_particles(wrapper, init = 1, min = 0, max = 1, add_options, samples,
  max_iter = 10, max_particles = 32768, ...)
}
\arguments{
\item{wrapper}{\code{\link{libbi}} (which has been run) to study}

\item{init}{initial number of particles (default: 1)}

\item{min}{minimum acceptance rate}

\item{max}{maximum acceptance rate}

\item{add_options}{list of additional options}

\item{samples}{number of samples to generate each iteration}

\item{max_iter}{maximum of iterations (default: 10)}

\item{max_particles}{maximum number of particles}

\item{...}{parameters for libbi$run}
}
\value{
a \code{\link{libbi}} with the desired proposal distribution
}
\description{
This function takes the provided \code{\link{libbi}} and
  runs MCMC at a single point (i.e., repeatedly proposing the same paramters),
  adapting the number of particles distribution until the desired
  acceptance rate is achieved. If a scale is given, it will be used
  to adapt the number of particles at each iteration.
}

