% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bi_model.R
\name{add_block}
\alias{add_block}
\alias{add_block.bi_model}
\title{Add a block to a LibBi model}
\usage{
\method{add_block}{bi_model}(x, name, lines, options, ...)
}
\arguments{
\item{x}{a \code{\link{bi_model}} object}

\item{name}{name of the block}

\item{lines}{character vector, lines in the block}

\item{options}{any options to the block}

\item{...}{ignored}
}
\value{
a \code{\link{bi_model}} object containing the new block
}
\description{
Add a block to a LibBi model. If that block exists, it will be removed first.
}

