% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libbi.R
\name{assert_files}
\alias{assert_files}
\alias{assert_files.libbi}
\title{Check that a LibBi wrapper has valid output}
\usage{
\method{assert_files}{libbi}(x, ...)
}
\arguments{
\item{x}{a \code{\link{libbi}} object}

\item{...}{ignored}
}
\value{
nothing (invisible NULL); an error will be thrown if there is a
  problem
}
\description{
This checks that the \code{\link{libbi}} object given has been run (via
  \code{\link{sample}}, \code{\link{filter}} or \code{\link{optimize}})) and
  the output file has not been modified since.
}
\keyword{internal}
