% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libbi.R
\name{attach}
\alias{attach}
\alias{attach.libbi}
\title{Attach a new file or data set to a \code{\link{libbi}} object}
\usage{
\method{attach}{libbi}(x, file, data, replace = FALSE, ...)
}
\arguments{
\item{x}{a \code{\link{libbi}} object}

\item{file}{the type of the file to attach, one of "output", "obs", "input" or "init"}

\item{data}{name of the file to attach, or a list of data frames that contain the outputs}

\item{...}{any options to \code{\link{bi_write}} (e.g., 'time_dim')}

\item{force}{attach the file even if one like this already exists in the libbi object}
}
\description{
Adds an (output, obs, etc.) file to a \code{\link{libbi}} object. This is useful to recreate a \code{\link{libbi}} object from the model and output files of a previous run
}
\examples{
bi <- libbi(model = system.file(package="rbi", "PZ.bi"))
example_output_file <- system.file(package="rbi", "example_output.nc")
bi <- attach(bi, "output", example_output_file)
}
