% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libbi.R
\name{attach_file}
\alias{attach_file}
\alias{attach_file.libbi}
\title{Deprecated (use 'attach' instead). Attach a new file or data set to a
  \code{\link{libbi}} object}
\usage{
\method{attach_file}{libbi}(x, file, data, force = FALSE, ...)
}
\arguments{
\item{x}{a \code{\link{libbi}} object}

\item{file}{the type of the file to attach, one of "output", "obs", "input"
or "init"}

\item{data}{name of the file to attach, or a list of data frames that contain
the outputs}

\item{force}{attach the file even if one like this already exists in the
libbi object}

\item{...}{any options to \code{\link{bi_write}} (e.g., 'time_dim')}
}
\value{
an updated \code{\link{libbi}} object
}
\description{
Deprecated (use 'attach' instead). Adds an (output, obs, etc.) file to a
  \code{\link{libbi}} object. This is useful to recreate a
  \code{\link{libbi}} object from the model and output files of a previous
  run
}
\keyword{internal}
