\name{bi-package}
\alias{bi-package}
\alias{bi}
\docType{package}
\title{
  bi
}
\description{
  R package to process the output of bi
}
\details{
\tabular{ll}{
Package: \tab bi\cr
Type: \tab Package\cr
Version: \tab 0.0-1\cr
Date: \tab 2012-06-18\cr
License: \tab \cr
LazyLoad: \tab no\cr
Depends: \tab ggplot2, ncdf\cr
}
}
\keyword{ package }
\author{
  Pierre E. Jacob <jacob@ceremade.dauphine.fr>
}

