% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bi_contents.R
\name{bi_contents}
\alias{bi_contents}
\title{Bi contents}
\usage{
bi_contents(read)
}
\arguments{
\item{read}{either a path to a NetCDF file, or a NetCDF connection created using \code{nc_open}, or a \code{\link{libbi}} object from which to read the output}
}
\value{
vector of variable names
}
\description{
This function gets the name of all the variables in the passed file, list or \code{\link{libbi}} object
}
\examples{
example_output_file <- system.file(package="rbi", "example_output.nc")
bi_contents(example_output_file)
}

