% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bi_dim_len.R
\name{bi_dim_len}
\alias{bi_dim_len}
\title{NetCDF dimension length}
\usage{
bi_dim_len(filename, dim)
}
\arguments{
\item{filename}{path to a NetCDF file}

\item{dim}{name of the dimension to check}
}
\value{
a number, the dimension length
}
\description{
This function returns the length of a dimension in a NetCDF file.
}
