% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bi_dim_values.R
\name{bi_dim_values}
\alias{bi_dim_values}
\title{NetCDF dimension values}
\usage{
bi_dim_values(filename, dim)
}
\arguments{
\item{filename}{path to a NetCDF file}

\item{dim}{name of the dimension to check}
}
\value{
dimension values
}
\description{
This function returns the values of a dimension in a NetCDF file.
}

