% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bi_file_ncdump.R
\name{bi_file_ncdump}
\alias{bi_file_ncdump}
\title{NetCDF File Print}
\usage{
bi_file_ncdump(filename)
}
\arguments{
\item{filename}{path to a NetCDF file}
}
\value{
None
}
\description{
This function prints the content of a file using the ncdump command line
}

