% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_dataset.R
\name{bi_generate_dataset}
\alias{bi_generate_dataset}
\title{Bi Generate Dataset}
\usage{
bi_generate_dataset(..., output_every = 1)
}
\arguments{
\item{...}{arguments to be passed to \code{\link{sample.libbi}}, especially
'model', 'end_time' and 'seed'.}

\item{output_every}{real; if given, \code{noutputs} will be set so that there
is output every \code{output_every} time steps; if set to 0, only generate
an output at the final time}
}
\value{
a \code{libbi} object, the generated data set
}
\description{
This function is deprecated and has been renamed to
\code{\link{generate_dataset}}
}
