% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bi_model.R
\name{bi_model_clone}
\alias{bi_model_clone}
\title{Clones a model (returning a new object with the same properties)}
\value{
cloned model
}
\description{
Returns a copy of the model with exactly the same content as the original model.
}
\examples{
model_file_name <- system.file(package="rbi", "PZ.bi")
PZ <- bi_model(filename = model_file_name)
PZ2 <- PZ$clone()
}
\seealso{
\code{\link{bi_model}}
}

