% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bi_model.R
\name{bi_model_delete}
\alias{bi_model_delete}
\title{Delete line(s) in a libbi model}
\arguments{
\item{num}{line number(s) to delete}
}
\value{
the updated bi model
}
\description{
Deletes one or more lines in a libbi model.
}
\examples{
model_file_name <- system.file(package="bi", "PZ.bi")
PZ <- bi_model(filename = model_file_name)
PZ$delete(2)
}
\seealso{
\code{\link{bi_model}}

Other bi_model.methods: \code{\link{bi_model_fix_noise}};
  \code{\link{bi_model_insert}};
  \code{\link{bi_model_propose_prior}};
  \code{\link{bi_model_update}};
  \code{\link{bi_model_write_model_file}};
  \code{\link{bi_model}}
}

