% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bi_model.R
\name{bi_model_delete_lines}
\alias{bi_model_delete_lines}
\title{Delete line(s) in a libbi model}
\arguments{
\item{num}{line number(s) to delete}
}
\value{
the updated bi model
}
\description{
Deletes one or more lines in a libbi model.
}
\examples{
model_file_name <- system.file(package="bi", "PZ.bi")
PZ <- bi_model(filename = model_file_name)
PZ$delete_line(2)
}
\seealso{
\code{\link{bi_model}}
}

