% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bi_model.R
\name{bi_model_fix_noise}
\alias{bi_model_fix_noise}
\title{Fix noise term of a libbi model}
\arguments{
\item{fixed}{a vector of values to be assigned to the (usually) noise value}

\item{default}{an optional default value to give to all noises not specified in \code{fixed}}
}
\value{
a bi model object of the new model
}
\description{
Replaces all noises with fixed values as given in 'fixed', otherwise as given in 'default'
}
\examples{
model_file_name <- system.file(package="bi", "PZ.bi")
PZ <- bi_model(filename = model_file_name)
PZ$fix_noise(fixed = c(alpha = 0))
}
\seealso{
\code{\link{bi_model}}
}

