% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bi_model.R
\name{bi_model_get_lines}
\alias{bi_model_get_lines}
\title{Get lines in a libbi model}
\arguments{
\item{spaces}{Number of spaces of indentation}
}
\value{
the libbi model as character vector
}
\description{
Gets a libbi model as character vector.
}
\examples{
model_file_name <- system.file(package="rbi", "PZ.bi")
PZ <- bi_model(filename = model_file_name)
PZ$get_lines()
}
\seealso{
\code{\link{bi_model}}
}

