% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bi_model.R
\name{bi_model_obs_to_noise}
\alias{bi_model_obs_to_noise}
\title{Copy observation variables into noise variables with "__sample_ prepended"}
\value{
a bi model object of the new model
}
\description{
This is used by libbi$run if \code{sample_obs} is set to TRUE
}
\examples{
model_file_name <- system.file(package="rbi", "PZ.bi")
PZ <- bi_model(filename = model_file_name)
PZ$obs_to_noise()
}
\seealso{
\code{\link{bi_model}}
}

