% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bi_model.R
\name{bi_model_remove_lines}
\alias{bi_model_remove_lines}
\title{Remove line(s) in a libbi model}
\arguments{
\item{num}{line number(s) to remove}
}
\value{
the updated bi model
}
\description{
Removes one or more lines in a libbi model.
}
\examples{
model_file_name <- system.file(package="bi", "PZ.bi")
PZ <- bi_model(filename = model_file_name)
PZ$remove_lines(2)
}
\seealso{
\code{\link{bi_model}}
}

