% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bi_model.R
\name{bi_model_replace_all}
\alias{bi_model_replace_all}
\title{replace all instances of one string with another}
\arguments{
\item{old}{the string to be replaced}

\item{new}{the string to replace the old string}
}
\value{
the updated bi model
}
\description{
Replace all instances of one string with another in a libbi model
}
\examples{
model_file_name <- system.file(package="rbi", "PZ.bi")
PZ <- bi_model(filename = model_file_name)
PZ$replace_all("alpha", "beta")
}
\seealso{
\code{\link{bi_model}}
}

