% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bi_model.R
\name{bi_model_set_name}
\alias{bi_model_set_name}
\title{Set the name of a bi model}
\arguments{
\item{name}{Name of the model}
}
\value{
the updated bi model
}
\description{
Changes the name of a bi model (first line of the .bi file) to the specified name.
}
\examples{
model_file_name <- system.file(package="rbi", "PZ.bi")
PZ <- bi_model(filename = model_file_name)
PZ$set_name("new_PZ")
}
\seealso{
\code{\link{bi_model}}
}

