% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bi_model.R
\name{bi_model_update_lines}
\alias{bi_model_update_lines}
\title{Update line(s) in a libbi model}
\arguments{
\item{num}{line number(s) to update}

\item{lines}{vector of iline(s)}
}
\value{
the updated bi model
}
\description{
Updates one or more lines in a libbi model.
}
\examples{
model_file_name <- system.file(package="bi", "PZ.bi")
PZ <- bi_model(filename = model_file_name)
PZ$update_lines(23, "alpha ~ normal(mu, sigma)")
}
\seealso{
\code{\link{bi_model}}
}

