% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bi_model.R
\name{bi_model_write_model_file}
\alias{bi_model_write_model_file}
\title{Writes a bi model to a file.}
\arguments{
\item{filename}{name of the file to be written}
}
\value{
the return value of the \code{\link{writeLines}} call.
}
\description{
Writes a bi model to a file given by \code{filename}. The extension '.bi' will be added if necessary.
}
\examples{
model_file_name <- system.file(package="bi", "PZ.bi")
PZ <- bi_model(filename = model_file_name)
PZ$write_model_file("PZ")
}
\seealso{
\code{\link{bi_model}}
}

