% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bi_write.R
\name{check_sparse_var}
\alias{check_sparse_var}
\title{Check if a variable is sparse}
\usage{
check_sparse_var(x, coord_cols, value_column)
}
\arguments{
\item{x}{data.table}

\item{coord_cols}{character vector of coordinate columns}

\item{value_column}{the name of the value column}
}
\value{
TRUE if the variable is sparse, FALSE otherwise
}
\description{
Check if a variable is sparse
}
\details{
Takes a data.table with given coordinate columns and a value column and
checks if all combinations of the coordinate columns are present for each
combination of the other columns.
}
\author{
Sebastian Funk
}
\keyword{internal}
