% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bi_model.R
\name{clean_model}
\alias{clean_model}
\title{Strip model code to its bare bones}
\usage{
clean_model(x)
}
\arguments{
\item{x}{a \code{\link{bi_model}} object}
}
\value{
a \code{\link{bi_model}}
}
\description{
Cleans the model by working out correct indents, removing long comments and
  merging lines
}
\seealso{
\code{\link{bi_model}}
}
\keyword{internal}

