% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libbi.R
\name{clone.libbi}
\alias{clone.libbi}
\title{Create a clone of a \code{\link{libbi}} object}
\usage{
\method{clone}{libbi}(x, ...)
}
\arguments{
\item{x}{the object to clone}

\item{...}{any additional arguments for \code{\link{run.libbi}}}
}
\value{
copy of the given object
}
\description{
This creates new versions of all the (input, output, etc.) files and creates a new object that is a copy of the given object.
}
\author{
Sebastian Funk
}
