% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bi_write.R
\name{create_coord_var}
\alias{create_coord_var}
\title{Create a coordinate variable}
\usage{
create_coord_var(
  name,
  dims,
  dim_factors,
  coord_dim,
  index_table,
  ns_dim,
  time_dim,
  nr_column,
  value_column
)
}
\arguments{
\item{name}{Name of the variable}

\item{dims}{Dimensions of all variables}

\item{dim_factors}{Factors of all dimensions}

\item{coord_dim}{Coordinate dimension}

\item{index_table}{data.table with index columns}

\item{ns_dim}{ns dimension}

\item{time_dim}{time dimension}

\item{nr_column}{nr column}

\item{value_column}{value column}
}
\value{
a list with information on the coordinate variable
}
\description{
Create a coordinate variable
}
\details{
Creates a coordinate variable with associated dimensions
}
\author{
Sebastian Funk
}
\keyword{internal}
