% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libbi.R
\name{create_working_folder}
\alias{create_working_folder}
\title{Internal function to create a temporary working folder}
\usage{
create_working_folder(x)
}
\arguments{
\item{x}{a \code{\link{libbi}} object}
}
\value{
a \code{\link{libbi}} object with updated working folder
}
\description{
The folder will be removed when the object is destroyed
}
\keyword{internal}
