% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bi_model.R
\name{Equals.bi_model}
\alias{Equals.bi_model}
\alias{==.bi_model}
\alias{`==.bi_model`}
\title{Check if two models are equal}
\usage{
\method{==}{bi_model}(e1, e2, ...)
}
\arguments{
\item{e1}{a \code{\link{bi_model}}}

\item{e2}{a \code{\link{bi_model}}}

\item{...}{ignored}
}
\value{
TRUE or FALSE, depending on whether the models are equal or not
}
\description{
Ignores differences in the model name.
}
\examples{
model_file_name <- system.file(package="rbi", "PZ.bi")
PZ <- bi_model(filename = model_file_name)
PZ == PZ # TRUE
}
