% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bi_model.R
\name{Extract_assign.bi_model}
\alias{Extract_assign.bi_model}
\alias{[<-.bi_model}
\alias{`[<-.bi_model`}
\title{Subset and replace model lines}
\usage{
\method{[}{bi_model}(x, i, ...) <- value
}
\arguments{
\item{x}{A bi_model}

\item{i}{A vector of line numbers}

\item{...}{ignored}

\item{value}{A vector of the same length as \code{i}, containing the
replacement strings}
}
\value{
the updated \code{bi_model} object
}
\description{
Extracts a subset of lines from the model and assigns new character strings.
}
\examples{
model_file_name <- system.file(package="rbi", "PZ.bi")
PZ <- bi_model(filename = model_file_name)
PZ[3:4] <- c("const e = 0.4", "const m_l = 0.05")
}
