% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bi_write.R
\name{factorise}
\alias{factorise}
\title{Convert character columns to factors in data}
\usage{
factorise(variables, levels)
}
\arguments{
\item{variables}{a \code{list} object, the names of which should be the
variable names and values should be either single values or data frames}

\item{levels}{factor levels, as a named list, each representing one column}
}
\value{
the \code{variables} argument with factorised columns
}
\description{
Convert character columns to factors in data
}
\author{
Sebastian Funk
}
\keyword{internal}
