% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bi_model.R
\name{find_block}
\alias{find_block}
\alias{find_block.bi_model}
\title{Find a block in a LibBi model}
\usage{
\method{find_block}{bi_model}(x, name)
}
\arguments{
\item{x}{a \code{\link{bi_model}} object}

\item{name}{of the block to find}
}
\value{
range of line numbers
}
\description{
Finds a block and returns the range of line numbers encompassed by that block.
}
\seealso{
\code{\link{bi_model}}
}
\keyword{internal}
