% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flatten.r
\name{flatten}
\alias{flatten}
\title{Flatten list of data frames
This function takes a list of data frames (such as, for example, returned by \code{\link{bi_read}}) and converts it to a flat data frame}
\usage{
flatten(x)
}
\arguments{
\item{x}{The list of data frames}
}
\value{
a data frame containing the flattened data
}
\description{
Flatten list of data frames
This function takes a list of data frames (such as, for example, returned by \code{\link{bi_read}}) and converts it to a flat data frame
}
