% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_generate_seed.R
\name{generate_seed}
\alias{generate_seed}
\title{Generate a random seed}
\usage{
generate_seed()
}
\value{
an integer, the random seed
}
\description{
This function generates a random seed using the random number generator of R.
}

