% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bi_write.R
\name{get_char_levels}
\alias{get_char_levels}
\title{Get the factor levels of all character columns in data}
\usage{
get_char_levels(...)
}
\arguments{
\item{...}{variable lists}
}
\value{
a list with elements that represent the factor levels present in
  character columns
}
\description{
Get the factor levels of all character columns in data
}
\author{
Sebastian Funk
}
\keyword{internal}
