% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bi_model.R
\name{get_const}
\alias{get_const}
\title{Get constants in a LibBi model}
\usage{
get_const(model)
}
\arguments{
\item{model}{a \code{\link{bi_model}} object}
}
\value{
list of constants (as names) and their values
}
\description{
Get constants contained in a LibBi model and their values. This will attempt to evaluate any calculation on the right hand side. Failing that, it will be returned verbatim.
}
