% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bi_model.R
\name{get_dims}
\alias{get_dims}
\title{Get dimensions in a LibBi model}
\usage{
get_dims(model, vars, type)
}
\arguments{
\item{vars}{a character vector of variable names; if given, only these variables names will be considered}

\item{type}{a character vector of one or more types}

\item{x}{a \code{\link{bi_model}} object}
}
\value{
list of dimensions (as names) and their sizes
}
\description{
Get dimensions contained in a LibBi model and their sizes
}
