% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bi_model.R
\name{get_name}
\alias{get_name}
\alias{get_name.bi_model}
\title{Get the name of a bi model}
\usage{
\method{get_name}{bi_model}(x, ...)
}
\arguments{
\item{x}{a \code{\link{bi_model}} object}

\item{...}{ignored}
}
\value{
the name of the model
}
\description{
Extracts the name of a bi model (first line of the .bi file).
}
\examples{
model_file_name <- system.file(package="rbi", "PZ.bi")
PZ <- bi_model(filename = model_file_name)
get_name(PZ)
}
\seealso{
\code{\link{bi_model}}
}

