% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/util_get_traces.R
\name{get_traces}
\alias{get_traces}
\title{Get the parameter traces}
\usage{
get_traces(run, all = FALSE, model, ...)
}
\arguments{
\item{run}{a \code{\link{libbi}} which has been run, or a file (in
which case either 'all' must be TRUE or a model given}

\item{all}{whether all variables in the run file should be
considered (otherwise, just parameters)}

\item{model}{a model to get the parameter names from; not needed if
'run' is given as a \code{\link{libbi}} object or 'all' is set to
TRUE}

\item{...}{parameters to \code{bi_read} (e.g., dimensions)}
}
\value{
data frame with parameter traces; this can be fed to \code{coda} routines
}
\description{
This function takes the provided \code{\link{libbi}} which has been
run and returns a data frame with the parameter traces.
}

