\name{getEvidence}
\alias{getEvidence}
\title{
  getEvidence
}
\description{
  Retrieve evidence from files.
}
\usage{
  getEvidence(files)
}
\arguments{
  \item{files}{
    Input files. Either one or a vector of strings.
  }
}
\details{
  The function returns a data frame containing the evidences.
}
\value{
  A pretty explicit data frame.
}

\seealso{
  \code{\link{bi}}
}
