% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installed_libbi_version.r
\name{installed_libbi_version}
\alias{installed_libbi_version}
\title{Get the LibBi version}
\usage{
installed_libbi_version(...)
}
\arguments{
\item{...}{any arguments for \code{\link{locate_libbi}} (esp.
\code{path_to_libbi})}
}
\value{
a character string with the installed version
}
\description{
This tries to find the libbi executable to determine the installed version
of LibBi
}
\author{
Sebastian Funk
}
\keyword{internal}
