% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bi_model.R
\name{is.empty}
\alias{is.empty}
\title{Check if a model is empty}
\arguments{
\item{x}{a \code{\link{bi_model}} object}

\item{...}{ignored}
}
\value{
the updated bi model
}
\description{
Checks if a model is empty (i.e., has been initialised without any content)
}
\examples{
model_file_name <- system.file(package="rbi", "PZ.bi")
PZ <- bi_model(filename = model_file_name)
is.empty(PZ)
}
\seealso{
\code{\link{bi_model}}
}

