% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bi_model.R
\name{is_empty}
\alias{is_empty}
\title{Check if a model is empty}
\usage{
is_empty(x)
}
\arguments{
\item{x}{a \code{\link{bi_model}} object}
}
\value{
the updated bi model
}
\description{
Checks if a model is empty (i.e., has been initialised without any content)
}
\seealso{
\code{\link{bi_model}}
}
\keyword{internal}
