% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libbi.R
\name{join}
\alias{join}
\alias{join.libbi}
\title{Join multiple \code{\link{libbi}} objects}
\usage{
\method{join}{libbi}(x, ...)
}
\arguments{
\item{x}{a \code{\link{libbi}} object}

\item{...}{ignored}
}
\description{
This function can be used to join multiple \code{\link{libbi}} objects into one (e.g., parallel MCMC runs into one long change)
}
