\name{launcher}
\alias{launcher}
\title{
  launcher
}
\description{
  Launch bi using the specified settings and additional arguments.
}
\usage{
  launcher(settings, Args)
}
\arguments{
  \item{settings}{
  }
  \item{Args}{
  }
}
\details{
}
\value{
}

\seealso{
  \code{\link{bi}},   \code{\link{settings}}
}
