% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_log2normw.R
\name{log2normw}
\alias{log2normw}
\title{Normalize log weights}
\usage{
log2normw(lw)
}
\arguments{
\item{lw}{a vector of real values}
}
\value{
a vector of normalized values (summing to 1)
}
\description{
This function takes a vector of real values,
then takes the exponential and divides by the sum.
Subtracting the max of the original values
increases the numerical stability.
}

