% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libbi.R
\name{logLik}
\alias{logLik}
\alias{logLik.libbi}
\title{Using the LibBi wrapper to logLik}
\usage{
\method{logLik}{libbi}(object, ...)
}
\arguments{
\item{object}{a \code{\link{libbi}} object}

\item{...}{options to be passed to \code{\link{run.libbi}}}
}
\value{
a vector of log-likelihood
}
\description{
The method \code{logLik} extracts the log-likelihood of a \code{libbi} object. This can be done, for example, after a call to \code{\link[rbi]{sample}} to inspect the chain log-likelihoods.

For the help page of the base R \code{logLik} function, see \code{\link[stats]{logLik}}.
}
