% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_merge_by_name.R
\name{merge_by_name}
\alias{merge_by_name}
\title{Merge lists by name}
\usage{
merge_by_name(...)
}
\arguments{
\item{...}{the lists to be merged}
}
\value{
the merged list
}
\description{
This function is used to create a single list from any number of named lists. Every name is unique in the new list. If names occur in multiple of the lists passed, later options overwrite earlier ones
}

