% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libbi.R
\name{optimise}
\alias{optimise}
\alias{optimise.libbi}
\title{Using the LibBi wrapper to optimise}
\usage{
\method{optimise}{libbi}(x, ...)
}
\arguments{
\item{x}{a \code{\link{libbi}} object}

\item{...}{options to be passed to \link{\code{run}}}
}
\description{
The method \code{optimise} of an instance of \code{\link{libbi}}
allows to launch \code{libbi} to optimise the parameters with respect to the likelihood or posterior distribution. See the options to \link{\code{run}} for how to specify the various components of sampling with LibBi, and the LibBi manual for all options that can be passed when the client is \code{optimise}.
}

