% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{option_list}
\alias{option_list}
\title{Convert string to option list}
\usage{
option_list(...)
}
\arguments{
\item{...}{any number of strings to convert}
}
\value{
a list of options and values
}
\description{
This function is used to convert an option string into a list of options. If a list is given, it will be kept as is
}
