% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{option_string}
\alias{option_string}
\title{Convert Options}
\usage{
option_string(...)
}
\arguments{
\item{...}{any number of lists of options, or strings (which will be left
unmodified). If lists are given, later arguments will override earlier ones}
}
\description{
This function is used to convert a list of options into an options string. If
  a string is given,  it will be taken as such.
}
