% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/util_output_to_proposal.R
\name{output_to_proposal}
\alias{output_to_proposal}
\title{Construct a proposal from run results}
\usage{
output_to_proposal(wrapper, scale)
}
\arguments{
\item{wrapper}{a \code{\link{libbi}} which has been run}

\item{scale}{a factor by which to scale all the standard deviations}
}
\value{
the updated bi model
}
\description{
This function takes the provided \code{\link{libbi}} which has been
run and returns a new model which has the proposal constructed from
the sample mean and standard deviation.
}

