% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bi_model.R
\name{print.bi_model}
\alias{print.bi_model}
\title{Print the lines of a LibBi model}
\usage{
\method{print}{bi_model}(x, ...)
}
\arguments{
\item{x}{a \code{\link{bi_model}} object}

\item{...}{ignored}
}
\description{
Prints all lines in a LibBi model
}

