% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bi_model.R, R/libbi.R
\name{print.bi_model}
\alias{print}
\alias{print.bi_model}
\alias{print.libbi}
\title{Print the lines of a LibBi model}
\usage{
\method{print}{bi_model}(x, spaces = 2, ...)

\method{print}{libbi}(x, verbose = FALSE, ...)
}
\arguments{
\item{x}{a \code{\link{bi_model}} object}

\item{spaces}{number of spaces for indentation}

\item{...}{ignored}

\item{verbose}{logical; if TRUE, locations of files and working folder should be printed}

\item{x}{a \code{\link{libbi}} object}

\item{...}{ignored}
}
\description{
Prints all lines in a LibBi model

This prints the model name, basic information such as number of iterations
  and timesteps run, as well as a list of variables.
}
\keyword{internal}

