% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libbi.R
\name{print_log}
\alias{print_log}
\title{Print the log file a \code{\link{libbi}} object}
\usage{
print_log(x)
}
\arguments{
\item{x}{a \code{\link{libbi}} object, or the name of the log file of a \code{\link{libbi}} run.}
}
\description{
This is useful for diagnosis after a \code{\link{libbi}} run
}
