% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bi_model.R
\name{propose_prior}
\alias{propose_prior}
\alias{propose_prior.bi_model}
\title{Propose from the prior in a libbi model}
\usage{
\method{propose_prior}{bi_model}(x)
}
\value{
a bi model object of the new model
}
\description{
Generates a version of the model where the proposal blocks are replaced by the prior blocks. This is useful for exploration of the likelihood surface.
}
\examples{
model_file_name <- system.file(package="rbi", "PZ.bi")
PZ <- bi_model(filename = model_file_name)
propose_prior(PZ)
}
\seealso{
\code{\link{bi_model}}
}

