% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libbi.R
\name{read_libbi}
\alias{read_libbi}
\title{Read results of a \code{LibBi} run from an RDS file or from a folder. This completely reconstructs the saved \code{LibBi} object}
\usage{
read_libbi(file, folder, ...)
}
\arguments{
\item{file}{name of the RDS file to read}

\item{folder}{name of the folder from which to read. This option cannot be used in conjunction
with \code{file}.}

\item{...}{any extra options to pass to \code{\link{read_libbi}} when creating the new object}
}
\value{
a \code{\link{libbi}} object
}
\description{
This reads all options, files and outputs of a \code{LibBi} run from a specified RDS file or
folder (if \code{split = TRUE} has been used with \code{save_libbi}).
}
