% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libbi.R
\name{read_results}
\alias{read_results}
\alias{read_results.libbi}
\title{Read results of a \code{LibBi} run from an RDS file. This completely reconstructs the saved \code{LibBi} object}
\usage{
\method{read_results}{libbi}(x, file, use_cache, ...)
}
\arguments{
\item{x}{a \code{\link{libbi}} object}

\item{file}{name of the RDS file to read}

\item{use_cache}{logical; whether to use the cache (default: \code{\link{libbi}} default)}

\item{...}{any options to \code{\link{libbi}}}
}
\value{
a \code{\link{libbi}} object
}
\description{
This reads all options, files and outputs of a \code{LibBi} run to an RDS file specified
}

