% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_var_input.R
\name{read_var_input}
\alias{read_var_input}
\title{Read variable from NetCDF file.}
\usage{
read_var_input(nc, name, coord, ps, ts, ...)
}
\arguments{
\item{nc}{NetCDF file handle}

\item{name}{Name of the variable}

\item{coord}{(optional) Demsions index.}

\item{ps}{(optional) Path indices.}

\item{ts}{(optional) Time indices.}

\item{...}{options for ncvar_get}
}
\value{
read values
}
\description{
Read variable from NetCDF file.
}
\author{
Lawrence Murray, \email{lawrence.murray@csiro.au}
}

