% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libbi.R
\name{readRDS}
\alias{readRDS}
\alias{readRDS.libbi}
\title{Read results of a \code{LibBi} run from an RDS file. This completely reconstructs the saved \code{LibBi} object}
\arguments{
\item{x}{a \code{\link{libbi}} object}

\item{file}{name of the RDS file to read}

\item{use_cache}{logical; whether to use the cache (default: \code{\link{libbi}} default)}

\item{...}{any options to \code{\link[base]{readRDS}}}
}
\value{
a \code{\link{libbi}} object
}
\description{
This reads all options, files and outputs of a \code{LibBi} run to an RDS file specified

For the help page of the base R \code{readRDS} function, see \code{\link[base]{readRDS}}.
}

