% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bi_model.R
\name{remove_vars}
\alias{remove_vars}
\title{Remove variables}
\usage{
remove_vars(x, vars)
}
\arguments{
\item{x}{a \code{\link{bi_model}} object}

\item{vars}{vector of variables to remove}
}
\value{
a bi model object of the new model

the updated \code{bi_model} object
}
\description{
Removes variables from the left-hand side of a model

Used by \code{\link{fix}} and \code{\link{to_input}}
}
\seealso{
\code{\link{bi_model}}
}
