% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bi_model.R
\name{rename}
\alias{rename}
\alias{rename.bi_model}
\title{Set the name of a bi model}
\usage{
\method{rename}{bi_model}(x, name)
}
\arguments{
\item{x}{a \code{\link{bi_model}} object}

\item{name}{Name of the model}
}
\value{
the updated bi model
}
\description{
Changes the name of a bi model (first line of the .bi file) to the specified name.
}
\examples{
model_file_name <- system.file(package="rbi", "PZ.bi")
PZ <- bi_model(filename = model_file_name)
PZ <- set_name(PZ, "new_PZ")
}
\seealso{
\code{\link{bi_model}}
}

