% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bi_model.R
\name{replace_all}
\alias{replace_all}
\alias{replace_all.bi_model}
\title{Replace all instances of a string with another in a model}
\usage{
\method{replace_all}{bi_model}(x, from, to, ...)
}
\arguments{
\item{x}{a \code{\link{bi_model}} object}

\item{from}{string to be replaced (a regular expression)}

\item{to}{new string (which can refer to the regular expression given as
\code{from})}

\item{...}{ignored}
}
\value{
the updated \code{bi_model} object
}
\description{
Takes every occurrence of one string and replaces it with another
}
\seealso{
\code{\link{bi_model}}
}
