% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libbi.R
\name{run}
\alias{run}
\title{Using the LibBi wrapper to launch LibBi}
\usage{
run(.Object, ...)
}
\arguments{
\item{...}{any unrecognised options will be added to \code{options}}

\item{x}{a \code{\link{libbi}} object}

\item{client}{client to pass to LibBi}

\item{options}{list of additional arguments to pass to the call to \code{LibBi}}

\item{config}{path to a configuration file, containing multiple arguments}

\item{add_options}{deprecated, replaced by \code{options}}

\item{log_file_name}{path to a file to text file to report the output of \code{LibBi}}

\item{stdoutput_file_name}{deprecated; use log_file_name instead}

\item{init}{initialisation of the model, either supplied as a list of values and/or data frames, or a (netcdf) file name, or a \code{\link{libbi}} object which has been run (in which case the output of that run is used as input)}

\item{input}{input of the model, either supplied as a list of values and/or data frames, or a (netcdf) file name, or a \code{\link{libbi}} object which has been run (in which case the output of that run is used as input)}

\item{obs}{observations of the model, either supplied as a list of values and/or data frames, or a (netcdf) file name, or a \code{\link{libbi}} object which has been run (in which case the output of that run is used as observations)}

\item{time_dim}{The time dimension in any R objects that have been passed (\code{init}, \code{input}) and \code{obs}); if not given, will be guessed}

\item{working_folder}{path to a folder from which to run \code{LibBi}; default to a temporary folder.}

\item{sample_obs}{logical; if set to TRUE, will sample observations}
}
\value{
a \code{\link{libbi}} object, except if \code{client} is 'rewrite',  in which case a \code{\link{bi_model}} object will be returned
}
\description{
The method \code{run} of an instance of \code{\link{libbi}}
allows to launch \code{LibBi} with a particular set of command line
arguments. Normally, this function would not be run by the user, but instead one of the client functions \code{\link{sample}}, \code{\link{filter}}, or \code{\link{optimise}}, or \code{\link{rewrite}}, which pass any options on to \code{run}. Note that any options specified here are stored in the \code{\link{libbi}} object and do not have to be specified again if another command is run on the object.
}
\examples{
bi_object <- libbi(model = system.file(package="rbi", "PZ.bi"))
\dontrun{run(bi_object, options=list(sample="smc2", nthreads = 1), verbose = TRUE)}
if (bi_object$run_flag) {
  bi_file_summary(bi_object$output_file_name)
}
}
\seealso{
\code{\link{libbi}}
}

