% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libbi.R
\name{sample}
\alias{sample}
\alias{sample.libbi}
\title{Using the LibBi wrapper to sample}
\usage{
\method{sample}{libbi}(x, ...)
}
\arguments{
\item{x}{a \code{\link{libbi}} object}

\item{...}{options to be passed to \link{\code{run}}}
}
\description{
The method \code{sample} of an instance of \code{\link{libbi}}
allows to launch \code{libbi} to sample from a (prior, posterior or joint) distribution. See the options to \link{\code{run}} for how to specify the various components of sampling with LibBi, and the LibBi manual for all options that can be passed when the client is \code{sample}.
}

