% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libbi.R
\name{sample_obs}
\alias{sample_obs}
\title{Sample observations from a LibBi model that has been run}
\usage{
sample_obs(x)
}
\arguments{
\item{x}{a \code{\link{libbi}} object}
}
\value{
the original \code{\link{libbi}} object with added variables in the output file for sampled observations
}
\description{
Sample observations from a LibBi model that has been run
}
\author{
Sebastian Funk
}
