% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/util_sample_observations.R
\name{sample_observations}
\alias{sample_observations}
\title{Sample observations from trajectories of a libbi model, one at each time step}
\usage{
sample_observations(wrapper, output_file_name)
}
\arguments{
\item{wrapper}{a \code{\link{libbi}} that has been run}

\item{output_file_name}{the name of the output file to write}
}
\description{
Sample observations from trajectories of a libbi model, one at each time step
}

