% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libbi.R
\name{save_libbi}
\alias{save_libbi}
\alias{save_libbi.libbi}
\title{Write results of a \code{LibBi} run to an RDS file}
\usage{
\method{save_libbi}{libbi}(x, filename, supplement, ...)
}
\arguments{
\item{x}{a \code{\link{libbi}} object}

\item{filename}{name of the RDS file to save to}

\item{supplement}{any supplementary data to save}

\item{...}{any options to \code{\link{saveRDS}}}
}
\description{
This saves all options, files and outputs of a \code{LibBi} run to an RDS file specified
}
