% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libbi.R
\name{saveRDS.libbi}
\alias{saveRDS.libbi}
\title{Write results of a \code{LibBi} run to an RDS file}
\arguments{
\item{x}{a \code{\link{libbi}} object}

\item{filename}{name of the RDS file to save to}

\item{...}{any options to \code{\link[base]{saveRDS}}}
}
\description{
This saves all options, files and outputs of a \code{LibBi} run to an RDS file specified

For the help page of the base R \code{saveRDS} function, see \code{\link[base]{saveRDS}}.
}

