% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libbi.R
\name{simulate}
\alias{simulate}
\alias{simulate.libbi}
\alias{simulate.bi_model}
\title{Using the LibBi wrapper to simulate}
\usage{
\method{simulate}{libbi}(x, ...)

\method{simulate}{bi_model}(x, ...)
}
\arguments{
\item{x}{a \code{\link{libbi}} or \code{\link{bi_model}} object, or the name of a file containing the model}

\item{...}{options to be passed to \code{\link{run.libbi}}}
}
\value{
a \code{\link{bi_model}} object
}
\description{
The method \code{simulate} launches \code{LibBi} to simulate a model by passing `target="joint"` to \code{LibBi}

If \code{x} is given as a 'bi_model', a \code{\link{libbi}} object will be created from the model
}
