% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libbi.R
\name{summary.libbi}
\alias{summary.libbi}
\title{Print summary information about a \code{\link{libbi}} object}
\arguments{
\item{object}{a \code{\link{libbi}} object}

\item{...}{ignored}
}
\description{
This reads in the output file of the \code{\link{libbi}} object (which has been run before) and prints summary information of parameters
}

