% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bi_model.R
\name{`[<-`}
\alias{`[<-`}
\title{Subset assignment to mode lines}
\arguments{
\item{x}{A bi_model}

\item{i}{A vector of line numbers}

\item{value}{A vector of the same length as \code{i}, containing the
replacement strings}
}
\description{
Assigns new character strings a subset of lines from the model
}
\examples{
model_file_name <- system.file(package="rbi", "PZ.bi")
PZ <- bi_model(filename = model_file_name)
PZ[3:4] <- c("const e = 0.4", "const m_l = 0.05")
}

