% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libbi.R
\name{update}
\alias{update}
\alias{update.libbi}
\title{Update a libbi object}
\usage{
\method{update}{libbi}(x, ...)
}
\arguments{
\item{x}{a \code{\link{libbi}} object}

\item{...}{ignored}
}
\value{
a \code{\link{libbi}} object with updated timestamps
}
\description{
This updates all the time stamps in a libbi object; it is useful after
(input, output, etc.) files have been changed outside the object itself.
}
