% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bi_model.R
\name{write}
\alias{write}
\alias{write.bi_model}
\title{Writes a bi model to a file.}
\usage{
\method{write}{bi_model}(x, filename)
}
\arguments{
\item{filename}{name of the file to be written}
}
\value{
the return value of the \code{\link{writeLines}} call.
}
\description{
Writes a bi model to a file given by \code{filename}. The extension '.bi' will be added if necessary.
}
\examples{
model_file_name <- system.file(package="rbi", "PZ.bi")
PZ <- bi_model(filename = model_file_name)
PZ$write("PZ")
}
\seealso{
\code{\link{bi_model}}
}

