% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/util_adapt_mcmc.R
\name{adapt_mcmc}
\alias{adapt_mcmc}
\title{Adapt the proposal distribution of MCMC using the covariance
  of samples}
\usage{
adapt_mcmc(wrapper, min = 0, max = 1, scale = 1, add_options, samples,
  max_iter = 10, ...)
}
\arguments{
\item{wrapper}{\code{link{libbi}} (which has been run) to study}

\item{min}{minimum acceptance rate}

\item{max}{maximum acceptance rate}

\item{scale}{scale multiplier/divider for the proposal. If >1 this
will be inverted.}

\item{add_options}{list of additional options}

\item{samples}{number of samples to generate each iteration}

\item{max_iter}{maximum of iterations (default: 10)}

\item{...}{parameters for libbi$run}
}
\value{
a \code{\link{libbi}} with the desired proposal distribution
}
\description{
This function takes the provided \code{\link{libbi}} object and
  runs MCMC, adapting the proposal distribution until the desired
  acceptance rate is achieved. If a scale is given, it will be used
  to adapt the proposal at each iteration
}

