% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libbi.R
\name{assert_output}
\alias{assert_output}
\alias{assert_output.libbi}
\title{Check that a LibBi wrapper has valid output}
\arguments{
\item{x}{a \code{\link{libbi}} object}

\item{...}{ignored}
}
\description{
This checks that the \code{\link{libbi}} object given has been run (via \code{\link{sample}}, \code{\link{filter}} or \code{\link{optimize}})) and the output file has not been modified since.
}
\keyword{internal}

