\name{bi-package}
\alias{bi-package}
\alias{bi}
\docType{package}
\title{
  bi
}
\description{
  R interface to bi
}
\details{

The package requires the software bi (Bayesian Inference) to be installed on the system.
Please follow the instructions on [URL] to install bi on your system.

Then have a look at the demos, a list of which is available through 
\code{demo(package="bi")}. For instance start by running \code{demo(package="launchPZ")}.

\tabular{ll}{
Package: \tab bi\cr
Type: \tab Package\cr
Version: \tab 0.0-1\cr
Date: \tab 2012-06-18\cr
License: \tab \cr
LazyLoad: \tab no\cr
Depends: \tab ggplot2, ncdf\cr
}
}
\keyword{ package }
\author{
  Pierre E. Jacob <jacob@ceremade.dauphine.fr>
}

