% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bi_generate_dataset.R
\name{bi_generate_dataset}
\alias{bi_generate_dataset}
\title{Bi Generate Dataset}
\usage{
bi_generate_dataset(..., seed)
}
\arguments{
\item{...}{arguments to be passed to \code{\link{libbi}} and \code{\link{sample}}, especially 'model' and 'end_time'.}

\item{seed}{random seed; see the seed option of \code{\link{sample}} for details.}
}
\value{
generated data set
}
\description{
This is a wrapper around \code{libbi sample --target joint --nsamples 1}, to generate a
synthetic dataset from a model. Parameters can be passed via the 'init' option
(see \code{\link{run.libbi}}, otherwise they are generated from the prior specified
in the model. The end time should be specified using the "end_time" option. If this is not given,
only a parameter set is sampled.
}

