% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bi_generate_dataset.R
\name{bi_generate_dataset}
\alias{bi_generate_dataset}
\title{Bi Generate Dataset}
\usage{
bi_generate_dataset(endtime, noutputs, ...)
}
\arguments{
\item{endtime}{final time index, so that data is generated from time 0 to time "endtime".}

\item{noutputs}{number of output points to be extracted from the hidden process; default is noutputs = endtime.}

\item{...}{arguments to be passed to \code{\link{libbi}} (with run = TRUE), especially 'model'}
}
\value{
path to the output file.
}
\description{
This is a wrapper around \code{libbi sample --target joint}, to generate
synthetic dataset from a model. Parameters can be passed via the 'init' option
(see \code{\link{libbi_run}}, otherwise they are generated from the prior specified
in the model.
}

