% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bi_write.R
\name{bi_init_file}
\alias{bi_init_file}
\title{Create init files for LibBi, retained for backwards compatibility}
\usage{
bi_init_file(...)
}
\arguments{
\item{...}{parameters passed to \code{bi_write}}
}
\value{
whatever \code{bi_write} returns
}
\description{
Users should use \code{bi_write} instead
}
\seealso{
\code{\link{bi_write}}
}

