% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bi_model.R
\name{bi_model}
\alias{bi_model}
\title{Bi Model}
\arguments{
\item{filename}{is the file name of the model file}
}
\description{
\code{bi_model} creates a model object for \code{Rbi} from a libbi file.
Once the instance is created, the model can either be fed to a \code{bi_wrapper}
object.
}
\examples{
bi_sir <- bi_model$new(filename = "sir.bi")
}
\seealso{
\code{\link{bi_model_fix_noise}}, \code{\link{bi_model_propose_prior}}, \code{\link{insert_lines}}, \code{\link{update_lines}}, \code{\link{delet_lines}}, \code{\link{write_model_file}},
#'

Other bi_model.methods: \code{\link{bi_model_delete}};
  \code{\link{bi_model_fix_noise}};
  \code{\link{bi_model_insert}};
  \code{\link{bi_model_propose_prior}};
  \code{\link{bi_model_update}};
  \code{\link{bi_model_write_model_file}}
}

