% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bi_model.R
\name{bi_model}
\alias{bi_model}
\title{Bi Model}
\arguments{
\item{filename}{is the file name of the model file}
}
\description{
\code{bi_model} creates a model object for \code{Rbi} from a libbi file.
Once the instance is created, the model can either be fed to a \code{\link{libbi}}
object.
}
\examples{
model_file_name <- system.file(package="rbi", "PZ.bi")
PZ <- bi_model(filename = model_file_name)
}
\seealso{
\code{\link{bi_model_fix}}, \code{\link{bi_model_propose_prior}}, \code{\link{bi_model_insert_lines}}, \code{\link{bi_model_update_lines}}, \code{\link{bi_model_remove_lines}}, \code{\link{bi_model_write_model_file}},
}

