% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bi_model.R
\name{bi_model_propose_prior}
\alias{bi_model_propose_prior}
\title{Propose from the prior in a libbi model}
\value{
a bi model object of the new model
}
\description{
Generates a version of the model where the proposal blocks are replaced by the prior blocks. This is useful for exploration of the likelihood surface.
}
\examples{
model_file_name <- system.file(package="bi", "PZ.bi")
PZ <- bi_model(filename = model_file_name)
PZ$propose_prior()
}
\seealso{
\code{\link{bi_model}}

Other bi_model.methods: \code{\link{bi_model_delete}};
  \code{\link{bi_model_fix_noise}};
  \code{\link{bi_model_insert}};
  \code{\link{bi_model_update}};
  \code{\link{bi_model_write_model_file}};
  \code{\link{bi_model}}
}

