% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bi_open.R
\name{bi_open}
\alias{bi_open}
\title{Bi open}
\usage{
bi_open(read)
}
\arguments{
\item{read}{either a path to a NetCDF file, or a NetCDF connection created using \code{nc_open}, or a \code{\link{libbi}} object from which to read the output}
}
\value{
open NetCDF connection
}
\description{
This function opens an NetCDF file
The file can be specified as a string to the filepath, in which
case a NetCDF connection is opened, or directly as a NetCDF connection.
}

