% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bi_read_file.R
\name{bi_read_file}
\alias{bi_read_file}
\title{Bi Read File}
\usage{
bi_read_file(file, dims, missval.threshold, variables, time_dim)
}
\arguments{
\item{file}{either a path to a NetCDF file, or a NetCDF connection created using \code{nc_open}}

\item{dims}{factors for dimensions}

\item{missval.threshold}{upper threshold for the likelihood}

\item{variables}{only extract given variables (for space saving)}

\item{time_dim}{name of time dimension (if any)}

\item{quiet}{suppress progress bar}
}
\value{
list of results
}
\description{
This function reads all variable from a NetCDF file.
The file can be specified as a string to the filepath, in which
case a NetCDF connection is opened, or directly as a NetCDF connection.
}

