% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bi_model.R
\name{Extract.bi_model}
\alias{Extract.bi_model}
\alias{[.bi_model}
\alias{`[.bi_model`}
\title{Subset model lines}
\usage{
\method{[}{bi_model}(x, i, ...)
}
\arguments{
\item{x}{A bi_model}

\item{i}{A vector of line numbers}

\item{...}{ignored}
}
\description{
Extracts a subset of lines from the model.
}
\examples{
model_file_name <- system.file(package="rbi", "PZ.bi")
PZ <- bi_model(filename = model_file_name)
PZ[3:4]
}
