% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_sample.R
\name{extract_sample}
\alias{extract_sample}
\title{Extract a sample from a \code{LibBi} run.}
\usage{
extract_sample(x, np, ...)
}
\arguments{
\item{x}{a \code{\link{libbi}} object which has been run, or a
list of data frames containing parameter traces (as returned by
from \code{bi_read}); if it is not a \code{\link{libbi}}
object, either 'all' must be TRUE or a model given}

\item{np}{iteration to extract; if set to "last", the last sample will be extracted. If not given a random sample will be extracted}

\item{...}{parameters to \code{bi_read} (e.g., dimensions)}
}
\value{
list of parameters and trajectories
}
\description{
This function takes the provided \code{\link{libbi}}
    object which has been run and extracts a data frame.
    traces.
}

