% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bi_model.R
\name{get_block}
\alias{get_block}
\alias{get_block.bi_model}
\title{Get the contents of a block in a LibBi model}
\usage{
\method{get_block}{bi_model}(x, name, ...)
}
\arguments{
\item{x}{a \code{\link{bi_model}} object}

\item{name}{name of the block}

\item{...}{ignored}
}
\value{
a character vector of the lines in the block
}
\description{
Returns the contents of a block in a LibBi model as a character vector of
  lines.
}
