% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bi_model.R
\name{get_dims}
\alias{get_dims}
\title{Get dimensions in a LibBi model}
\usage{
get_dims(model, type)
}
\arguments{
\item{model}{a \code{\link{bi_model}} object}

\item{type}{a character vector of one or more types}
}
\value{
list of dimensions (as names) and their sizes
}
\description{
Get dimensions contained in a LibBi model and their sizes
}
