% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bi_model.R
\name{get_lines}
\alias{get_lines}
\title{Get lines in a libbi model}
\usage{
get_lines(x, spaces = 2)
}
\arguments{
\item{x}{a \code{\link{bi_model}} object}

\item{spaces}{Number of spaces of indentation}
}
\value{
the libbi model as character vector
}
\description{
Gets a libbi model as character vector.
}
\seealso{
\code{\link{bi_model}}, \code{\link{`[.bi_model`}}
}
\keyword{internal}

