% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/util_get_traces.R
\name{get_traces}
\alias{get_traces}
\title{Get the parameter traces}
\usage{
get_traces(wrapper, ...)
}
\arguments{
\item{wrapper}{a \code{bi_wrapper} which has been run}

\item{...}{parameters to \code{bi_read_file} (e.g., dimensions)}
}
\value{
the updated bi model
}
\description{
This function takes the provided \code{bi_wrapper} which has been
run and returns a data frame with the parameter traces.
}

