% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bi_model.R
\name{is_empty}
\alias{is_empty}
\alias{is_empty.bi_model}
\title{Check if a model is empty}
\usage{
is_empty.bi_model(x, name, ...)
}
\arguments{
\item{x}{a \code{\link{bi_model}} object}

\item{...}{ignored}
}
\value{
the updated bi model
}
\description{
Checks if a model is empty (i.e., has been initialised without any content)
}
\seealso{
\code{\link{bi_model}}
}
\keyword{internal}

