% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libbi.R
\name{libbi_run}
\alias{bi_run}
\alias{libbi}
\alias{libbi_run}
\title{Using the LibBi wrapper to launch LibBi}
\arguments{
\item{add_options}{additional arguments to pass to the call to \code{libbi}}

\item{stdoutput_file_name}{path to a file to text file to report the output of \code{libbi}}

\item{init}{initialisation of the model, either supplied as a list of values and/or data frames, or a (netcdf) file name, or a \code{\link{libbi}} object which has been run (in which case the output of that run is used as input)}

\item{input}{input of the model, either supplied as a list of values and/or data frames, or a (netcdf) file name, or a \code{\link{libbi}} object which has been run (in which case the output of that run is used as input)}

\item{obs}{observations of the model, either supplied as a list of values and/or data frames, or a (netcdf) file name, or a \code{\link{libbi}} object which has been run (in which case the output of that run is used as observations)}

\item{verbose}{if TRUE, will run libbi with the '--verbose' option}

\item{...}{any onrecognised options will be added to \code{add_options}}
}
\value{
a list containing the absolute paths to the results; it is stored in the
\code{result} field of the instance of \code{\link{libbi}}.
}
\description{
The method \code{run} of an instance of \code{\link{libbi}}
allows to launch \code{libbi} with a particular set of command line
arguments.
}
\examples{
bi_object <- libbi$new(client = "sample",
                       model = system.file(package="bi", "PZ.bi"),
                       global_options = list(sampler = "smc2"))
bi_object$run(add_options=list(nthreads = 1), verbose = TRUE)
bi_file_summary(bi_object$result$output_file_name)
}
\seealso{
\code{\link{libbi}}
}

