% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/util_nc_create_from_list.R
\name{netcdf_create_from_list}
\alias{netcdf_create_from_list}
\title{Create NetCDF File from R list}
\usage{
netcdf_create_from_list(filename, variables)
}
\arguments{
\item{filename}{a path to a NetCDF file to write the variable into, which will be overwritten
if it already exists.}

\item{variables}{a \code{list}}
}
\value{
None, but creates a NetCDF file at the specified path.
}
\description{
This function creates a NetCDF file given a list.
}
\details{
The list must follow the following rules. Each element of the list
must itself be one of:

1) a list with two keys; the first key must be named "values" and
contains a numeric vector; the second key must be named "dimension"
and contains a string giving the dimension name.

2) a data frame with a "value" column and any number of other
columns indicating one or more dimensions

3) a numeric vector of length one, with no dimensions

The name of the list elements itself is used to create the
corresponding variable in the NetCDF file.
}
\note{
Two elements of the given list can possibly have the same
  dimension name.
}
\examples{
filename <- tempfile(pattern="dummy", fileext=".nc")
a <- list(values = 1:3, dimension = "dim_a")
b <- list(values = 1:5, dimension = "dim_b")
c <- list(values = 5:9, dimension = "dim_b")
d <- 3
e <- data.frame(dim_a = rep(1:3, time = 2), dim_c = rep(1:2, each = 3), value = 1:6)
variables <- list(a=a, b=b, c=c, d=d, e=e)
netcdf_create_from_list(filename, variables)
bi_file_ncdump(filename)
}

