% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bi_model.R
\name{obs_to_noise}
\alias{obs_to_noise}
\title{Copy obs variables to state variables (with '__sample_' prepended)}
\usage{
obs_to_noise(x)
}
\arguments{
\item{x}{a \code{\link{bi_model}} object}
}
\value{
a \code{\link{bi_model}}
}
\seealso{
\code{\link{bi_model}}
}
\keyword{internal}

