% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libbi.R
\name{predict}
\alias{predict}
\alias{predict.libbi}
\title{Using the LibBi wrapper to predict}
\usage{
\method{predict}{libbi}(x, ...)
}
\arguments{
\item{x}{a \code{\link{libbi}} object}

\item{...}{any arguments to be passed to \code{\link{sample}}}
}
\value{
an updated \code{\link{libbi}} object
}
\description{
The method \code{predict} is an alias for \code{sample(target="prediction")}.
  Usually, an \code{init} object or file should be given containing posterior
  samples.

For the help page of the base R \code{optimise} function, see
  \code{\link[stats]{optimise}}.
}
