% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libbi.R
\name{print}
\alias{print}
\alias{print.libbi}
\title{Print information about a \code{\link{libbi}} object}
\usage{
\method{print}{libbi}(x, verbose = FALSE, ...)
}
\arguments{
\item{x}{a \code{\link{libbi}} object}

\item{verbose}{logical; if TRUE, locations of files and working folder should be printed}

\item{...}{ignored}
}
\description{
This prints the model name, basic information such as number of iterations
  and timesteps run, as well as a list of variables.
}

