% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bi_model.R
\name{propose_prior}
\alias{propose_prior}
\title{Propose from the prior in a libbi model}
\arguments{
\item{x}{a \code{\link{bi_model}} object}

\item{...}{ignored}
}
\value{
a bi model object of the new model
}
\description{
Generates a version of the model where the proposal blocks are replaced by the prior blocks. This is useful for exploration of the likelihood surface.
}
\seealso{
\code{\link{bi_model}}
}
\keyword{internal}

