% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libbi.R
\name{read_libbi}
\alias{read_libbi}
\title{Read results of a \code{LibBi} run from an RDS file. This completely reconstructs the saved \code{LibBi} object}
\usage{
read_libbi(file, ...)
}
\arguments{
\item{file}{name of the RDS file to read}

\item{...}{any extra options to pass to \code{\link{read_libbi}} when creating the new object}
}
\value{
a \code{\link{libbi}} object
}
\description{
This reads all options, files and outputs of a \code{LibBi} run to an RDS file specified
}

