% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bi_model.R
\name{remove}
\alias{remove}
\alias{remove.bi_model}
\title{Remove line(s) in a libbi model}
\usage{
\method{remove}{bi_model}(x, num)
}
\arguments{
\item{num}{line number(s) to remove}
}
\value{
the updated bi model
}
\description{
Removes one or more lines in a libbi model.
}
\examples{
model_file_name <- system.file(package="rbi", "PZ.bi")
PZ <- bi_model(filename = model_file_name)
PZ <- remove_lines(PZ, 2)
}
\seealso{
\code{\link{bi_model}}
}

