% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bi_model.R
\name{replace_lines}
\alias{replace_lines}
\alias{replace_lines.bi_model}
\title{Update line(s) in a libbi model}
\usage{
replace_lines.bi_model(x, num, lines, ...)
}
\arguments{
\item{x}{a \code{\link{bi_model}} object}

\item{num}{line number(s) to update}

\item{lines}{vector of line(s)}

\item{...}{ignored}
}
\value{
the updated bi model
}
\description{
Updates one or more lines in a libbi model.
}
\seealso{
\code{\link{bi_model}}
}
\keyword{internal}

