% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libbi.R
\name{rewrite}
\alias{rewrite}
\alias{rewrite.libbi}
\title{Using the LibBi wrapper to rewrite}
\usage{
\method{rewrite}{libbi}(x, ...)
}
\arguments{
\item{...}{options to be passed to \link{\code{run}}}
}
\description{
The method \code{rewrite} of an instance of \code{\link{libbi}}
allows to launch \code{LibBi} to rewrite a model to inspect its internal representation in \code{LibBi}
#' @param x a \code{\link{libbi}} object
}

