% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libbi.R
\name{rewrite}
\alias{rewrite}
\alias{rewrite.libbi}
\alias{rewrite}
\alias{rewrite.bi_model}
\title{Using the LibBi wrapper to rewrite}
\usage{
\method{rewrite}{libbi}(x, ...)

\method{rewrite}{bi_model}(x, ...)
}
\arguments{
\item{x}{a \code{\link{libbi} or \link{bi_model}} object, or the name of a file containing the model}

\item{...}{options to be passed to \code{\link{run.libbi}}}
}
\value{
a \code{\link{bi_model}} object
}
\description{
The method \code{rewrite} launches \code{LibBi} to rewrite a model to inspect its internal representation in \code{LibBi}

If \code{x} is given as a 'bi_model', a \code{\link{libbi}} object will be created from the model
}
