% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libbi.R
\name{sample_obs}
\alias{sample_obs}
\title{Sample observations from a LibBi model that has been run}
\usage{
sample_obs(x, ...)
}
\arguments{
\item{x}{a \code{\link{libbi}} object}

\item{...}{any arguments to be passed to \code{\link{predict}}}
}
\value{
a \code{\link{libbi}} object with sampled observations
}
\description{
Sample observations from a LibBi model that has been run
}
\author{
Sebastian Funk
}
