% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libbi.R
\name{save_libbi}
\alias{save_libbi}
\alias{save_libbi.libbi}
\title{Write results of a \code{LibBi} run to an RDS file}
\usage{
\method{save_libbi}{libbi}(x, name, supplement, split = FALSE, ...)
}
\arguments{
\item{x}{a \code{\link{libbi}} object}

\item{name}{name of the RDS file(s) to save to. If \code{split=TRUE}, this will be taken as a base for the names of the files to be created, e.g. 'dir/name.rds' to create files of the form name_....rds in directory 'dir'.}

\item{supplement}{any supplementary data to save}

\item{split}{Logical, defaults to \code{FALSE}. Should the objects from the
\code{LibBi} run be saved separately in a folder.}

\item{...}{any options to \code{\link{saveRDS}}}
}
\description{
This saves all options, files and outputs of a \code{LibBi} run to an RDS file specified
}
