% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libbi.R
\name{summary}
\alias{summary}
\alias{summary.libbi}
\title{Print summary information about a \code{\link{libbi}} object}
\usage{
\method{summary}{libbi}(object, type = c("param", "state", "noise",
  "obs"), ...)
}
\arguments{
\item{object}{a \code{\link{libbi}} object}

\item{type}{one of "param" (default), "state", "noise" or "obs", the variable type to summarise}

\item{...}{ignored}
}
\description{
This reads in the output file of the \code{\link{libbi}} object (which has been run before) and prints summary information of parameters
}
