% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bi_model.R
\name{to_input}
\alias{to_input}
\title{Convert variables into inputs}
\usage{
to_input(x, vars)
}
\arguments{
\item{x}{a \code{\link{bi_model}} object}

\item{vars}{vector of variables to convert to inputs}
}
\value{
a bi model object of the new model
}
\description{
Used for predictions, if one doesn't want to re-simulate state/noise trajectories
}
\seealso{
\code{\link{bi_model}}
}
\keyword{internal}
