% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bi_model.R
\name{var_names}
\alias{var_names}
\title{Get variables}
\arguments{
\item{x}{a \code{\link{bi_model}} object}

\item{type}{a character vector of one or more types}

\item{dim}{logical; if set to TRUE, names will contain dimensions in brackets}

\item{opt}{logical; if set to TRUE, names will contain options (e.g., has_output)}

\item{...}{ignored}
}
\value{
variable names
}
\description{
Get all variable names of one or more type(s)
}
\details{
This returns all variable names of a certain type ("param", "state", "obs", "noise", "const") contained in the model of a \code{\link{libbi}} object
}

